+++ src/main/java/br/com/zalf/prolog/webservice/entrega/ImportResource.java
@@ -2,7 +2,7 @@ package br.com.zalf.prolog.webservice.entrega;
 
 import br.com.zalf.prolog.webservice.commons.network.Response;
 import br.com.zalf.prolog.webservice.commons.util.Log;
-import br.com.zalf.prolog.webservice.commons.util.Now;
+import br.com.zalf.prolog.webservice.commons.util.date.Now;
 import br.com.zalf.prolog.webservice.commons.util.Platform;
 import br.com.zalf.prolog.webservice.commons.util.UsedBy;
 import br.com.zalf.prolog.webservice.entrega.mapa.MapaService;