+++ src/main/java/br/com/zalf/prolog/webservice/entrega/escaladiaria/EscalaDiariaReader.java
@@ -1,6 +1,6 @@
 package br.com.zalf.prolog.webservice.entrega.escaladiaria;
 
-import br.com.zalf.prolog.webservice.commons.util.DateUtils;
+import br.com.zalf.prolog.webservice.commons.util.date.DateUtils;
 import br.com.zalf.prolog.webservice.commons.util.XlsxConverter;
 import com.univocity.parsers.csv.CsvParser;
 import com.univocity.parsers.csv.CsvParserSettings;