+++ src/main/java/br/com/zalf/prolog/webservice/frota/pneu/servico/relatorio/ServicoRelatorioService.java
@@ -21,8 +21,8 @@ public final class ServicoRelatorioService {
         try {
             return dao.getEstratificacaoServicosFechadosReport(
                     codUnidades,
-                    ProLogDateParser.validateAndParse(dataInicial),
-                    ProLogDateParser.validateAndParse(dataFinal));
+                    ProLogDateParser.toLocalDate(dataInicial),
+                    ProLogDateParser.toLocalDate(dataFinal));
         } catch (SQLException e) {
             Log.e(TAG, String.format("Erro ao buscar o relatório que estratifica os serviços fechados (REPORT). \n" +
                     "Unidades: %s \n" +
@@ -40,8 +40,8 @@ public final class ServicoRelatorioService {
             dao.getEstratificacaoServicosFechadosCsv(
                     outputStream,
                     codUnidades,
-                    ProLogDateParser.validateAndParse(dataInicial),
-                    ProLogDateParser.validateAndParse(dataFinal));
+                    ProLogDateParser.toLocalDate(dataInicial),
+                    ProLogDateParser.toLocalDate(dataFinal));
         } catch (SQLException | IOException e) {
             Log.e(TAG, String.format("Erro ao buscar o relatório que estratifica os serviços fechados (CSV). \n" +
                     "Unidades: %s \n" +
@@ -57,8 +57,8 @@ public final class ServicoRelatorioService {
         try {
             return dao.getEstratificacaoServicosAbertosReport(
                     codUnidades,
-                    ProLogDateParser.validateAndParse(dataInicial),
-                    ProLogDateParser.validateAndParse(dataFinal));
+                    ProLogDateParser.toLocalDate(dataInicial),
+                    ProLogDateParser.toLocalDate(dataFinal));
         } catch (SQLException e) {
             Log.e(TAG, String.format("Erro ao buscar o relatório que estratifica os serviços abertos (REPORT). \n" +
                     "Unidades: %s \n" +
@@ -76,8 +76,8 @@ public final class ServicoRelatorioService {
             dao.getEstratificacaoServicosAbertosCsv(
                     outputStream,
                     codUnidades,
-                    ProLogDateParser.validateAndParse(dataInicial),
-                    ProLogDateParser.validateAndParse(dataFinal));
+                    ProLogDateParser.toLocalDate(dataInicial),
+                    ProLogDateParser.toLocalDate(dataFinal));
         } catch (SQLException | IOException e) {
             Log.e(TAG, String.format("Erro ao buscar o relatório que estratifica os serviços abertos (CSV). \n" +
                     "Unidades: %s \n" +