+++ src/main/java/br/com/zalf/prolog/webservice/gente/controleintervalo/relatorios/ControleIntervaloRelatorioResource.java
@@ -2,6 +2,7 @@ package br.com.zalf.prolog.webservice.gente.controleintervalo.relatorios;
 
 import br.com.zalf.prolog.webservice.commons.report.Report;
 import br.com.zalf.prolog.webservice.commons.util.Required;
+import br.com.zalf.prolog.webservice.errorhandling.exception.ProLogException;
 import br.com.zalf.prolog.webservice.interceptors.auth.Secured;
 
 import javax.ws.rs.*;
@@ -17,7 +18,7 @@ import java.util.List;
 @Produces(MediaType.APPLICATION_JSON + ";charset=utf-8")
 public class ControleIntervaloRelatorioResource {
 
-    private ControleIntervalosRelatorioService service = new ControleIntervalosRelatorioService();
+    private ControleIntervaloRelatorioService service = new ControleIntervaloRelatorioService();
 
     @GET
     @Secured
@@ -125,7 +126,8 @@ public class ControleIntervaloRelatorioResource {
                                                             @PathParam("cpf") @Required String cpf,
                                                             @PathParam("codTipoIntervalo") @Required String codTipoIntervalo,
                                                             @QueryParam("dataInicial") @Required String dataInicial,
-                                                            @QueryParam("dataFinal") @Required String dataFinal) {
+                                                            @QueryParam("dataFinal") @Required String dataFinal)
+            throws ProLogException {
         return service.getFolhaPontoRelatorio(
                 codUnidade,
                 codTipoIntervalo,
@@ -172,8 +174,8 @@ public class ControleIntervaloRelatorioResource {
     @Path("/total-tempo-por-tipo-intervalo/{codUnidade}/{codTipoIntervalo}/csv")
     public StreamingOutput getTotalTempoByTipoIntervaloCsv(@PathParam("codUnidade") @Required final Long codUnidade,
                                                            @PathParam("codTipoIntervalo") @Required final String codTipoIntervalo,
-                                                           @QueryParam("dataInicial") @Required final String dataInicial,
-                                                           @QueryParam("dataFinal") @Required final String dataFinal) {
+                                                           @QueryParam("dataInicial") @Required String dataInicial,
+                                                           @QueryParam("dataFinal") @Required String dataFinal) {
         return outputStream -> service.getTotalTempoByTipoIntervaloCsv(
                 outputStream,
                 codUnidade,
@@ -185,10 +187,11 @@ public class ControleIntervaloRelatorioResource {
     @GET
     @Secured
     @Path("/total-tempo-por-tipo-intervalo/{codUnidade}/{codTipoIntervalo}/report")
-    public Report getTotalTempoByTipoIntervaloReport(@PathParam("codUnidade") @Required final Long codUnidade,
-                                                     @PathParam("codTipoIntervalo") @Required final String codTipoIntervalo,
-                                                     @QueryParam("dataInicial") @Required final String dataInicial,
-                                                     @QueryParam("dataFinal") @Required final String dataFinal) {
+    public Report getTotalTempoByTipoIntervaloReport(
+            @PathParam("codUnidade") @Required final Long codUnidade,
+            @PathParam("codTipoIntervalo") @Required final String codTipoIntervalo,
+            @QueryParam("dataInicial") @Required String dataInicial,
+            @QueryParam("dataFinal") @Required String dataFinal) throws Throwable {
         return service.getTotalTempoByTipoIntervaloReport(
                 codUnidade,
                 codTipoIntervalo,