+++ src/main/java/br/com/zalf/prolog/webservice/gente/controleintervalo/relatorios/ControleIntervaloRelatorioService.java
@@ -4,6 +4,8 @@ import br.com.zalf.prolog.webservice.Injection;
 import br.com.zalf.prolog.webservice.commons.report.Report;
 import br.com.zalf.prolog.webservice.commons.util.Log;
 import br.com.zalf.prolog.webservice.commons.util.ProLogDateParser;
+import br.com.zalf.prolog.webservice.errorhandling.exception.ProLogException;
+import br.com.zalf.prolog.webservice.errorhandling.exception.ProLogExceptionHandler;
 import org.jetbrains.annotations.NotNull;
 
 import java.io.IOException;
@@ -15,10 +17,11 @@ import java.util.List;
 /**
  * Created by Zart on 28/08/2017.
  */
-public class ControleIntervalosRelatorioService {
+public class ControleIntervaloRelatorioService {
 
+    private static final String TAG = ControleIntervaloRelatorioService.class.getSimpleName();
     private ControleIntervaloRelatoriosDao dao = Injection.provideControleIntervaloRelatoriosDao();
-    private static final String TAG = ControleIntervalosRelatorioService.class.getSimpleName();
+    private final ProLogExceptionHandler exceptionHandler = Injection.provideProLogExceptionHandler();
 
     public void getIntervalosCsv(OutputStream out, Long codUnidade, Long dataInicial, Long dataFinal, String cpf) {
         try {
@@ -129,8 +132,8 @@ public class ControleIntervalosRelatorioService {
                     codUnidade,
                     codTipoIntervalo,
                     cpf,
-                    ProLogDateParser.validateAndParse(dataInicial),
-                    ProLogDateParser.validateAndParse(dataFinal));
+                    ProLogDateParser.toLocalDate(dataInicial),
+                    ProLogDateParser.toLocalDate(dataFinal));
         } catch (SQLException | IOException e) {
             Log.e(TAG, String.format("Erro ao buscar o relatório csv no padrão da portaria 1510. \n" +
                     "codUnidade: %d \n" +
@@ -145,23 +148,23 @@ public class ControleIntervalosRelatorioService {
                                                             @NotNull final String codTipoIntervalo,
                                                             @NotNull final String cpf,
                                                             @NotNull final String dataInicial,
-                                                            @NotNull final String dataFinal) {
+                                                            @NotNull final String dataFinal) throws ProLogException {
         try {
-            final List<FolhaPontoRelatorio> folhaPontoRelatorio = dao.getFolhaPontoRelatorio(
+            return dao.getFolhaPontoRelatorio(
                     codUnidade,
                     codTipoIntervalo,
                     cpf,
-                    ProLogDateParser.validateAndParse(dataInicial),
-                    ProLogDateParser.validateAndParse(dataFinal));
-            return folhaPontoRelatorio;
-        } catch (SQLException e) {
-            Log.e(TAG, String.format("Erro ao buscar o relatório de folha de ponto. \n" +
+                    ProLogDateParser.toLocalDate(dataInicial),
+                    ProLogDateParser.toLocalDate(dataFinal));
+        } catch (final Throwable e) {
+            final String errorMessage = String.format("Erro ao buscar o relatório de folha de ponto. \n" +
                     "codUnidade: %d \n" +
                     "codTipoIntervalo: %s \n" +
                     "cpf: %s \n" +
                     "dataInicial: %s \n" +
-                    "dataFinal: %s", codUnidade, codTipoIntervalo, cpf, dataInicial, dataFinal), e);
-            throw new RuntimeException(e);
+                    "dataFinal: %s", codUnidade, codTipoIntervalo, cpf, dataInicial, dataFinal);
+            Log.e(TAG, errorMessage, e);
+            throw exceptionHandler.map(e, errorMessage);
         }
     }
 
@@ -174,8 +177,8 @@ public class ControleIntervalosRelatorioService {
             return dao.getMarcacoesComparandoEscalaDiariaReport(
                     codUnidade,
                     codTipoIntervalo,
-                    ProLogDateParser.validateAndParse(dataInicial),
-                    ProLogDateParser.validateAndParse(dataFinal));
+                    ProLogDateParser.toLocalDate(dataInicial),
+                    ProLogDateParser.toLocalDate(dataFinal));
         } catch (SQLException e) {
             Log.e(TAG, String.format("Erro ao buscar report do relatório de marcações comparando com escala diária. \n" +
                     "codUnidade: %d \n" +
@@ -196,8 +199,8 @@ public class ControleIntervalosRelatorioService {
                     out,
                     codUnidade,
                     codTipoIntervalo,
-                    ProLogDateParser.validateAndParse(dataInicial),
-                    ProLogDateParser.validateAndParse(dataFinal));
+                    ProLogDateParser.toLocalDate(dataInicial),
+                    ProLogDateParser.toLocalDate(dataFinal));
         } catch (IOException | SQLException e) {
             Log.e(TAG, String.format("Erro ao buscar csv do relatório de marcações comparando com escala diária. \n" +
                     "codUnidade: %d \n" +
@@ -218,15 +221,15 @@ public class ControleIntervalosRelatorioService {
                     out,
                     codUnidade,
                     codTipoIntervalo,
-                    ProLogDateParser.validateAndParse(dataInicial),
-                    ProLogDateParser.validateAndParse(dataFinal));
-        } catch (IOException | SQLException e) {
+                    ProLogDateParser.toLocalDate(dataInicial),
+                    ProLogDateParser.toLocalDate(dataFinal));
+        } catch (final Throwable t) {
             Log.e(TAG, String.format("Erro ao buscar csv do relatório de total de tempo para cada tipo de intervalo. \n" +
                     "codUnidade: %d \n" +
                     "codTipoIntervalo: %s \n" +
                     "dataInicial: %s \n" +
-                    "dataFinal: %s", codUnidade, codTipoIntervalo, dataInicial, dataFinal), e);
-            throw new RuntimeException(e);
+                    "dataFinal: %s", codUnidade, codTipoIntervalo, dataInicial, dataFinal), t);
+            throw new RuntimeException(t);
         }
     }
 
@@ -234,20 +237,22 @@ public class ControleIntervalosRelatorioService {
     public Report getTotalTempoByTipoIntervaloReport(@NotNull final Long codUnidade,
                                                      @NotNull final String codTipoIntervalo,
                                                      @NotNull final String dataInicial,
-                                                     @NotNull final String dataFinal) {
+                                                     @NotNull final String dataFinal) throws ProLogException {
         try {
             return dao.getTotalTempoByTipoIntervaloReport(
                     codUnidade,
                     codTipoIntervalo,
-                    ProLogDateParser.validateAndParse(dataInicial),
-                    ProLogDateParser.validateAndParse(dataFinal));
-        } catch (SQLException e) {
-            Log.e(TAG, String.format("Erro ao buscar report do relatório de total de tempo para cada tipo de intervalo. \n" +
-                    "codUnidade: %d \n" +
-                    "codTipoIntervalo: %s \n" +
-                    "dataInicial: %s \n" +
-                    "dataFinal: %s", codUnidade, codTipoIntervalo, dataInicial, dataFinal), e);
-            throw new RuntimeException(e);
+                    ProLogDateParser.toLocalDate(dataInicial),
+                    ProLogDateParser.toLocalDate(dataFinal));
+        } catch (Throwable e) {
+            final String errorMessage = String.format(
+                    "Erro ao buscar report do relatório de total de tempo para cada tipo de intervalo. \n" +
+                            "codUnidade: %d \n" +
+                            "codTipoIntervalo: %s \n" +
+                            "dataInicial: %s \n" +
+                            "dataFinal: %s", codUnidade, codTipoIntervalo, dataInicial, dataFinal);
+            Log.e(TAG, errorMessage, e);
+            throw exceptionHandler.map(e, errorMessage);
         }
     }
 }