+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Change Log
 
 ### Features
 * Cria validator para o insert do Modelo de Quiz
+* Cria relatório para mostrar o total de tempo gasto em cada tipo de intervalo por cada colaborador
 
 ### Refactors
 * Adiciona permissão de movimentação na busca de uma recapadora específica
@@ -58,7 +59,6 @@ Change Log
 ### Features
 * Cria back end para salvar mensagens de contato de possíveis clientes
 * Ativa método de listagem de marcações de um colaborador
-* Cria relatório para mostrar o total de tempo gasto em cada tipo de intervalo por cada colaborador
 
 ### Refactor
 * Altera componente da Dashboard para mostrar apenas os veículos ativos e não apenas os veículos com pneus vinculados